fix(desktop): render automation run settings as composer chip dropdowns
Access mode, model provider, model, and reasoning rendered as four native
selects, while the operator sets exactly these values everywhere else in
the app through the composer footer's chip dropdowns. Same settings,
different control — for no reason beyond the selects being what the
editor started with.

ComposerDropdown (and useDismissableMenu, which seven other composer
menus share) moves out of the 13k-line Composer.tsx into its own module,
verbatim, and the composer imports it from there. The automation editor
renders the four settings as one wrapping chip row — the same idiom as
the reply box — keeping the catalog wiring intact: provider changes drop
a model the new catalog does not list, model changes drop a reasoning
effort the new model does not offer, and stored values missing from the
catalog stay expressible.

Tests drive the chips the way a user does now (click the chip, click an
option) instead of firing change events at selects.
